Changes requested in review

+++ sys/kern/vm_physmem.c
@@ -214,12 +214,12 @@ vm_page_t *vm_page_alloc(size_t npages) {
   return pm_take_page(fl);
 }
 
-vm_page_t *vm_page_alloc_zero(int npages) {
+vm_page_t *vm_page_alloc_zero(size_t npages) {
   vm_page_t *pages = vm_page_alloc(npages);
   if (!pages)
     return NULL;
 
-  for (int i = 0; i < npages; i++)
+  for (size_t i = 0; i < npages; i++)
     pmap_zero_page(&pages[i]);
   return pages;
 }
